requirements.txt
openai>=1.46.0,<2.0.0
langchain>=0.2.11,<0.3.0
langchain-openai>=0.1.20,<0.2.0
PyYAML>=6.0.1,<7.0.0
PyGithub>=2.3.0,<3.0.0
